delete y update parametros ajustes en los permisos y seguridad

+++ src/app/components/admin/admin.component.ts
@@ -11,7 +11,10 @@ export class AdminComponent {
   constructor(private cookieService: CookieService, public router: Router) {}
 
   logout(){
-    this.cookieService.delete('token');            
-    this.router.navigateByUrl(''); 
+    if(confirm("Desea salir del sistema ?")) {
+      this.cookieService.delete('token');            
+      this.router.navigateByUrl('');   
+    }
+    
   }
 }
